Compiling and working with u16 MAPSEC
Plenty of changes; struct was being padded/aligned automatically, but the generated .inc files do not account for this change in struct size

+++ include/map_preview_screen.h
@@ -41,7 +41,7 @@ enum MapPreviewScreenId
 
 struct MapPreviewScreen
 {
-    u8 mapsec;
+    u16 mapsec;
     u8 type;
     u16 flagId;
     const void *tilesptr;
@@ -49,17 +49,17 @@ struct MapPreviewScreen
     const void *palptr;
 };
 
-u16 MapPreview_CreateMapNameWindow(u8 id);
+u16 MapPreview_CreateMapNameWindow(u16 id);
 void MapPreview_SetFlag(u16 a0);
-u16 MapPreview_GetDuration(u8 id);
-bool8 MapHasPreviewScreen(u8 mapsec, u8 type);
+u16 MapPreview_GetDuration(u16 id);
+bool8 MapHasPreviewScreen(u16 mapsec, u8 type);
 bool32 ForestMapPreviewScreenIsRunning(void);
-const struct MapPreviewScreen * GetDungeonMapPreviewScreenInfo(u8 mapsec);
-bool32 MapHasPreviewScreen_HandleQLState2(u8 mapsec, u8 type);
+const struct MapPreviewScreen * GetDungeonMapPreviewScreenInfo(u16 mapsec);
+bool32 MapHasPreviewScreen_HandleQLState2(u16 mapsec, u8 type);
 void MapPreview_InitBgs(void);
-void MapPreview_LoadGfx(u8 mapsec);
+void MapPreview_LoadGfx(u16 mapsec);
 bool32 MapPreview_IsGfxLoadFinished(void);
 void MapPreview_Unload(s32 windowId);
-void MapPreview_StartForestTransition(u8 mapsec);
+void MapPreview_StartForestTransition(u16 mapsec);
 
 #endif //GUARD_MAP_PREVIEW_SCREEN_H
